+++ lib/patches/digest/sha2.rb
@@ -1,0 +1,2 @@
+# Load the TruffleRuby specific sha2 file instead
+require_relative '../../../lib/mri/digest/sha2'